Moved payments into single file
Did not need a separate PAIN payments file, as payments implemented are
all pain currently.
As different payments get implemented they will fall under their own
packages
